requirements.txt
streamlit
anthropic
python-dotenv
pydantic
pandas
supabase
